Вся логика для событий написана, но проблемы с сериализацией/десериализацией string и local date time

+++ main-service/src/main/java/ru/practicum/dto/GetEventParametersUserRequest.java
@@ -17,7 +17,7 @@ public class GetEventParametersUserRequest {
 
     private String text;
 
-    private List<Integer> categories;
+    private List<Long> categories;
 
     private Boolean paid;
